delete restriction rules && courseHelpers
